feat: add nice search artist bar

+++ lib/data/api/artist/dtos/search_artist.dto.dart
@@ -9,7 +9,7 @@ class SearchArtistDto with _$SearchArtistDto {
     String? query,
     @Default(1) int page,
     @Default(10) int limit,
-    int? minRating,
+    double? minRating,
   }) = _SearchArtistDto;
 
   factory SearchArtistDto.fromJson(Map<String, dynamic> json) =>
